requirements.txt
streamlit
pdfplumber
gspread
google-auth
google-auth-oauthlib
pandas
